Simplify the public API of ExternalBindings
