+++ src/main/java/org/rumbledb/api/ExternalBindings.java
@@ -4,13 +4,14 @@ import lombok.Value;
 import lombok.experimental.Accessors;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
+import org.rumbledb.api.binding.Binding;
+import org.rumbledb.api.binding.DataFrameBinding;
+import org.rumbledb.api.binding.ItemSequenceBinding;
 import org.rumbledb.context.Name;
 
-import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 
 /**
  * Query-scoped external bindings supplied at execution time.
@@ -22,124 +23,34 @@ import java.util.Objects;
 @Value
 @Accessors(fluent = true)
 public class ExternalBindings {
-    private static final ExternalBindings EMPTY = new Builder().build();
-
     private Map<Name, Binding> variables;
 
-    private ExternalBindings(Map<Name, Binding> variables) {
-        this.variables = Collections.unmodifiableMap(new LinkedHashMap<>(variables));
+    public ExternalBindings() {
+        this.variables = new LinkedHashMap<>();
     }
 
-    public static Builder builder() {
-        return new Builder();
+    public void bindItem(String variableName, Item item) {
+        this.bindItems(variableName, List.of(item));
     }
 
-    public static ExternalBindings empty() {
-        return EMPTY;
+    public void bindItems(String variableName, List<Item> items) {
+        Name variable = Name.createVariableInNoNamespace(variableName);
+        if (this.variables.containsKey(variable)) {
+            throw new IllegalArgumentException("Variable " + variableName + " is already bound.");
+        }
+
+        this.variables.put(variable, new ItemSequenceBinding(items));
     }
 
-    public boolean isEmpty() {
-        return this.variables.isEmpty();
+    public void bindDataFrame(String variableName, Dataset<Row> dataFrame) {
+        Name variable = Name.createVariableInNoNamespace(variableName);
+        if (this.variables.containsKey(variable)) {
+            throw new IllegalArgumentException("Variable " + variableName + " is already bound.");
+        }
+        this.variables.put(variable, new DataFrameBinding(dataFrame));
     }
 
-    public interface Binding {
-        Kind kind();
-    }
-
-    public enum Kind {
-        ITEM_SEQUENCE,
-        DATAFRAME
-    }
-
-    @Value
-    public static class ItemSequenceBinding implements Binding {
-        List<Item> items;
-
-        public ItemSequenceBinding(List<Item> items) {
-            this.items = List.copyOf(Objects.requireNonNull(items, "items"));
-        }
-
-        @Override
-        public Kind kind() {
-            return Kind.ITEM_SEQUENCE;
-        }
-    }
-
-    @Value
-    public static class DataFrameBinding implements Binding {
-        Dataset<Row> dataFrame;
-
-        public DataFrameBinding(Dataset<Row> dataFrame) {
-            this.dataFrame = Objects.requireNonNull(dataFrame, "dataFrame");
-        }
-
-        @Override
-        public Kind kind() {
-            return Kind.DATAFRAME;
-        }
-    }
-
-    public static class Builder {
-        private final Map<Name, Binding> variables = new LinkedHashMap<>();
-
-        public Builder variables(Map<Name, Binding> variables) {
-            this.variables.clear();
-            this.variables.putAll(Objects.requireNonNull(variables, "variables"));
-            return this;
-        }
-
-        /**
-         * Convenience equivalent of binding a singleton sequence of one item.
-         */
-        public Builder bindItem(String variableName, Item item) {
-            return bindItem(Name.createVariableInNoNamespace(variableName), item);
-        }
-
-        /**
-         * Convenience equivalent of binding a singleton sequence of one item.
-         */
-        public Builder bindItem(Name variableName, Item item) {
-            return bindItems(variableName, List.of(Objects.requireNonNull(item, "item")));
-        }
-
-        /**
-         * Replaces {@code RumbleRuntimeConfiguration.setExternalVariableValue(String, List<Item>)}.
-         */
-        public Builder bindItems(String variableName, List<Item> items) {
-            return bindItems(Name.createVariableInNoNamespace(variableName), items);
-        }
-
-        /**
-         * Replaces {@code RumbleRuntimeConfiguration.setExternalVariableValue(Name, List<Item>)}.
-         */
-        public Builder bindItems(Name variableName, List<Item> items) {
-            this.variables.put(
-                Objects.requireNonNull(variableName, "variableName"),
-                new ItemSequenceBinding(items)
-            );
-            return this;
-        }
-
-        /**
-         * Replaces {@code RumbleRuntimeConfiguration.setExternalVariableValue(String, Dataset<Row>)}.
-         */
-        public Builder bindDataFrame(String variableName, Dataset<Row> dataFrame) {
-            return bindDataFrame(Name.createVariableInNoNamespace(variableName), dataFrame);
-        }
-
-        /**
-         * Replaces {@code RumbleRuntimeConfiguration.setExternalVariableValue(Name, Dataset<Row>)}.
-         */
-        public Builder bindDataFrame(Name variableName, Dataset<Row> dataFrame) {
-            this.variables.put(
-                Objects.requireNonNull(variableName, "variableName"),
-                new DataFrameBinding(dataFrame)
-            );
-            return this;
-        }
-
-        public ExternalBindings build() {
-            return new ExternalBindings(this.variables);
-        }
+    public void unbind(String variableName) {
+        this.variables.remove(Name.createVariableInNoNamespace(variableName));
     }
 }
