+++ src/main/java/org/rumbledb/api/binding/Binding.java
@@ -1,0 +1,10 @@
+package org.rumbledb.api.binding;
+
+public interface Binding {
+    public enum Kind {
+        ITEM_SEQUENCE,
+        DATAFRAME
+    }
+
+    Kind kind();
+}
